Add a model method to post praise to Mattermost.
The transparency group was looking for ways to gently remind folks of
the staff-thanks function, and one idea was to have occasional reminders of
recent praise postings. This would allow a praise object to ship its
details to Mattermost. Then we could perhaps have a job that sends the
latest praise to mattermost a couple times a week.